feat: Refactor fulltext search to use @appland/search
