+++ packages/cli/package.json
@@ -94,6 +94,7 @@
     "@appland/openapi": "workspace:^1.8.0",
     "@appland/rpc": "workspace:^1.13.0",
     "@appland/scanner": "workspace:^1.86.0",
+    "@appland/search": "workspace:^1.0.0",
     "@appland/sequence-diagram": "workspace:^1.12.0",
     "@octokit/rest": "^20.0.1",
     "@sidvind/better-ajv-errors": "^0.9.1",
